[HumanBenchmark][Frontend][Codebase] move sidebar into its own folder
