+++ front-end/src/App.jsx
@@ -1,17 +1,13 @@
-import { useState } from 'react'
-import HomePage from './HomePage/HomePage'
-import './App.css'
-
-
-
+import { useState } from "react";
+import HomePage from "./HomePage/HomePage";
+import "./App.css";
 
 function App() {
-
   return (
     <>
       <HomePage />
     </>
-  )
+  );
 }
 
-export default App
+export default App;
